requirement.txt
langchain
langchain-community
openai
python-dotenv
pytest
pytest-cov
